State Pattern Example
I chose the state pattern since it is very useful in a game development setting. A common example of where states are used is the player object. The player will have many different factors as to when, how, and why they do something. Separating functionality of the player into states makes it very easy to manage and define what the player is doing or can do at a given time or state.