Prototype Pattern
The prototype allows me to clone objects of the different types easily due to them having a common interface. In my code example, I have a bomb interface which each type of bomb implements. So in main, I can create objects of type bomb and then set them to be types of bomb or to be clones of other existing bomb objects.